Ajout d'un fichier de configuration pour le générateur de projet.

+++ Modele_Projet/type_C/Tester.sh
@@ -1,27 +1,41 @@
 #!/bin/bash
 : '
-	* AUTEUR : Erwan PECHON
-	* VERSION : 0.1
-	* DATE : Ven. 22 Sept. 2023 15:13:54
-	* Script qui sert à lancer les différent programmes de tests du projet.
+* AUTEUR : Erwan PECHON
+* VERSION : 0.1
+* DATE : Ven. 22 Sept. 2023 15:13:54
+* Script qui sert à lancer les différent programmes de tests du projet.
 '
 
 
 # CRÉATION DES VARIABLES GLOBALES
+dossierActu=`pwd`
+dossierProjet=`realpath "$0"`
+dossierProjet=`dirname "$dossierProjet"`
+
 nbParam_min=1
+
 val=1
+gdb=-1
 com=-1
 
+lstTest="for F in \$(ls $dossierProjet/bin/test_*[^.]\.out | tr \"\\\n\" \"\\\t\") ; do basename \"\$F\" ; done"
 
 # CONSIGNE D'UTILISATION
-Usage="Usage : $0 [-nc] [--no_valgrind] [--compile] <classe>..."
+Usage="Usage : $0 [-nc|-gc] [--no_valgrind|--debug] [--compile] <classe>..."
 Detail_noValgrind="\t- -n ou --no_valgrind : Si présent désactive valgrind."
+Detail_gdb="\t- -g ou --debug : Si présent désactive valgrind, et active gdb."
 Detail_compile="\t- -c ou --compile : Si présent, lance un 'make all' avant de tester les programmes."
 Detail_classe="\t- <classe> : Lance le programme test_<classe>, si il existe."
 Detail="$Usage\n-h : Affiche cette aide.\n$Detail_noValgrind\n$Detail_compile\n$Detail_classe"
-if ( test $1 == -h ) then
+if ( test "$1" == "-h" ) then
+	echo ""
+	echo "Affichage de l'aide :"
 	echo -e "$Detail"
-	exit 1
+	echo ""
+	echo -ne "Liste des classes actuellement disponible. (faite make all pour mettre à jour.) : \n\t>"
+	echo "$(eval "$lstTest")" | tr "\n" "\t"
+	echo ""
+	exit 0
 fi
 
 
@@ -45,40 +59,67 @@ while ( test $stop -eq 0 -a $# -lt $nbParam_min ) ; do
 	case $1 in
 		# Si grand paramètre :
 		"--no_valgrind") val=`expr $val * -1`; shift ;;
+		"--debug") gdb=`expr $gdb * -1`; shift ;;
 		"--compile") com=`expr $com * -1`; shift ;;
 		# Si deux des paramètres réduit :
 		"-nc")
-			val=`expr $val * -1`;
-			com=`expr $com * -1`;
-			shift ;;
-		"-cn")
-			com=`expr $com * -1`;
-			val=`expr $val * -1`;
-			shift ;;
+		val=`expr $val * -1`;
+		com=`expr $com * -1`;
+		shift ;;
+	"-cn")
+		com=`expr $com * -1`;
+		val=`expr $val * -1`;
+		shift ;;
+	"-gc")
+		gdb=`expr $gdb * -1`;
+		com=`expr $com * -1`;
+		shift ;;
+	"-cg")
+		gdb=`expr $gdb * -1`;
+		com=`expr $com * -1`;
+		shift ;;
 		# Si un seul des paramètres réduit :
 		"-n") val=`expr $val * -1`; shift ;;
+		"-n") gdb=`expr $gdb * -1`; shift ;;
 		"-c") com=`expr $com * -1`; shift ;;
 		# Sinon
 		*) stop = 1
-	esac
+esac
 done
 
-#Test des paramètres optionnels no_valgrind :
-if ( test $val -eq -1 ) then
+#Test des paramètres optionnels debug :
+if ( test "$gdb" -ne -1 ) then
 	val=0
+	gdb=1
+else #Test des paramètres optionnels no_valgrind :
+	if ( test "$val" -ne -1 ) then
+		val=1
+		gdb=0
+	else
+		val=0
+		gdb=0
+	fi
 fi
 
 #Test des paramètres optionnels compile :
-if ( test $com -eq -1 ) then
+if ( test "$com" -ne -1 ) then
+	com=1
+else
 	com=0
 fi
 
 #Lecture des paramètres classe :
 lstExec=()
-while ( test $# -ne 0 ) ; do
-	lstExec=( ${lstExec[*]} ./bin/test_$1 )
+execAll=-1
+if ( test $# -eq 1 -a "$1" == "*" ) then
+	execAll=1
 	shift
-done
+else
+	while ( test $# -ne 0 ) ; do
+		lstExec=( ${lstExec[*]} ./bin/test_$1 )
+		shift
+	done
+fi
 # ATTENTION : nombre de repetition de ce paramètre inconnue
 
 ##Test du paramètre !!!!! <?????> :
@@ -93,17 +134,24 @@ done
 
 # SCRIPT
 clear
-#Rafraichir les éxecutable à tester
-rm -vf ${lstExecutable[*]}
 #Compiler
 if ( test $com -eq 1 ) then
 	make clean all
 else
-	make ${lstExecutable[*]}
+	if ( test $execAll -eq -1 ) then
+		rm -vf ${lstExecutable[*]}
+		make ${lstExecutable[*]}
+	fi
+fi
+if ( test $execAll -ne -1 ) then
+	lstExec=()
+	for F in `echo "$(eval "$lstTest")" | tr "\n" "\t"` ; do
+		lstExec=( ${lstExec[*]} ./bin/test_$F )
+	done
 fi
 #Executer les programmes demandé.
-Fval="valgrind.log"
-Fval2="valgrind-2.log"
+Fval="$dossierProjet/valgrind.log"
+Fval2="$dossierProjet/valgrind-2.log"
 if (test $val -eq 1 ) then
 	rm -f $Fval $Fval2
 	touch $Fval
@@ -123,12 +171,17 @@ for File in ${lstExec[*]} ; do
 		echo "#########################################" >> $Fval
 		rm -f $Fval2
 	else
-		$File
+		if ( test $gdb -eq 1 ) then
+			gdb $File
+		else
+			$File
+		fi
 	fi
 done
 
 echo -e "\n\n\t\tFIN DES TESTS\n\n"
-ls ./bin/test_*[^.]? | cut -c12-
+echo -ne "Liste des classes actuellement disponible. (faite make all pour mettre à jour.) : \n\t>"
+echo "$(eval "$lstTest")" | tr "\n" "\t"
 echo -e "\n\n"
 ls
